HW4. Created new filmCards component
