+++ src/pages/my-list-screen/my-list-screen.tsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import FilmCard from '../../components/film-card/film-card';
+import FilmCards from '../../components/film-cards/film-cards';
 import Footer from '../../components/footer/footer';
 import HeaderUserBlock from '../../components/header-user-block/header-user-block';
 import Logo from '../../components/logo/logo';
@@ -10,17 +9,6 @@ type MyListProps = {
 }
 
 function MyListScreen({films}: MyListProps): JSX.Element {
-  const [hoveredFilmId, setHoveredFilmId] = React.useState<number | null>(null);
-
-
-  const handleMouseOver = (id: number) => {
-    setHoveredFilmId(id);
-  };
-
-  const handleMouseOut = () => {
-    setHoveredFilmId(null);
-  };
-
   return (
     <div className="user-page">
       <header className="page-header user-page__head">
@@ -33,9 +21,7 @@ function MyListScreen({films}: MyListProps): JSX.Element {
         <h2 className="catalog__title visually-hidden">Catalog</h2>
 
         <div className="catalog__films-list">
-          {
-            films.map((film) => <FilmCard key={film.id} film={film} isHovered={hoveredFilmId === film.id} mouseOverHandler={handleMouseOver} mouseOutHandler={handleMouseOut}/>)
-          }
+          <FilmCards films={films}/>
         </div>
       </section>
